make the code right running

+++ src/main/java/entity/GdNaviLinkTest.java
@@ -10,14 +10,14 @@ public class GdNaviLinkTest {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     private int pathID;
-    private Long distance;
-    private Long duration;
-    private Long speed;
+    private Double distance;
+    private Double duration;
+    private Double speed;
 
     public GdNaviLinkTest() {
         super();
     }
-    public GdNaviLinkTest(int pathID,Long distance,Long duration,Long speed){
+    public GdNaviLinkTest(int pathID,Double distance,Double duration,Double speed){
         this.pathID=pathID;
         this.distance=distance;
         this.duration=duration;
